Chatbot/Chatbot_project/requirements.txt
openai
gradio
request
python-dotenv
bs4
pypdf2
faiss